kafka-clients/requirements.txt
avro-python3==1.8.2
kafka==1.3.5

